chore(ui): move career and leaderboard screens
